Rename stack file to cdk_stack.py for brevity
